% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_codebook_syntax.R
\name{write_codebook_syntax}
\alias{write_codebook_syntax}
\title{Dynamically support CERP variable management}
\usage{
write_codebook_syntax(
  codebookWorkbookFilename,
  codebookWorkbookFileLocation,
  codebookSyntaxFileLocation
)
}
\arguments{
\item{codebookWorkbookFilename}{The full name of the finished codebook workbook for this evaluation dataset}

\item{codebookWorkbookFileLocation}{The location of the finished codebook workbook}

\item{codebookSyntaxFileLocation}{The location of where the output file will go}
}
\value{
An .xlsx file with columns to support the codebook management process in SPSS.
}
\description{
Dynamically support CERP variable management
}
\details{
Make an excel file to dynamically support CERP variable management
}
\examples{
\donttest{
write_codebook_syntax(
codebookWorkbookFilename = "GradCohortIDEALS_2021_pre_codebookWorkbook.xlsx",
codebookWorkbookFileLocation = "[path to codebookWorkbook file on Dropbox]",
codebookSyntaxFileLocation = "[path to output file on Dropbox]"
)
}

}
\author{
Evelyn Yarzebinski, Computing Research Association
}
