% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_codebook_workbook.R
\name{write_codebook_workbook}
\alias{write_codebook_workbook}
\title{Dynamically support CERP variable management}
\usage{
write_codebook_workbook(
  variableWorkbookFilename,
  variableWorkbookTabName,
  variableWorkbookFileLocation,
  dictionaryFilename,
  dictionaryFileLocation,
  codebookFilename,
  codebookFileLocation
)
}
\arguments{
\item{variableWorkbookFilename}{The full name of the finished variable workbook for this evaluation dataset}

\item{variableWorkbookTabName}{The name of the tab where the finished variable workbook content is}

\item{variableWorkbookFileLocation}{The location of the finished variable workbook}

\item{dictionaryFilename}{The name of the data dictionary exported from Qualtrics}

\item{dictionaryFileLocation}{The location of the data dictionary exported from Qualtrics}

\item{codebookFilename}{The name of the codebook workbook (the output of this script)}

\item{codebookFileLocation}{The location where the codebook workbook will be saved}
}
\value{
An .xlsx file with columns to support the codebook management process in SPSS.
}
\description{
Dynamically support CERP variable management
}
\details{
Make an excel file to dynamically support CERP variable management
}
\examples{
\dontrun{
write_codebook_workbook(
variableWorkbookFilename = "GradCohortIDEALS_2021_pre_variableManagement.xlsx",
variableWorkbookTabName = "varWorkbook",
variableWorkbookFileLocation = "[path to variable workbook on Dropbox]",
dictionaryFilename = "GCIDEALS2021_pre_codebook.csv",
dictionaryFileLocation = "[path to SPSS dictionary on Dropbox]",
codebookFilename = "GradCohortIDEALS_2021_codebookWorkbook.xlsx",
codebookFileLocation = "[path to where to write final codebook file]"
)
}

}
\author{
Evelyn Yarzebinski, Computing Research Association
}
