% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_variable_workbook.R
\name{write_variable_workbook}
\alias{write_variable_workbook}
\title{Dynamically support CERP variable management}
\usage{
write_variable_workbook(
  csvName,
  savName,
  rawDataLocation,
  variableWorkbookFilename,
  variableWorkbookFileLocation
)
}
\arguments{
\item{csvName}{The full, raw name of the csv file from Qualtrics}

\item{savName}{The full, raw name of the sav file from Qualtrics}

\item{rawDataLocation}{The pointer to the raw csv and sav file (both files must be in the same folder)}

\item{variableWorkbookFilename}{What you want to set as the name of your exported Excel file}

\item{variableWorkbookFileLocation}{Where you want the output file to be saved}
}
\value{
An .xlsx file with columns to support the variable management process in SPSS.
}
\description{
Dynamically support CERP variable management
}
\details{
Make an excel file to dynamically support CERP variable management
}
\examples{
\donttest{
write_variable_workbook(
csvName = "PRE.CISE Pre-survey_August 4, 2021_09.46.csv",
savName = "PRE.CISE Pre-survey_August 4, 2021_09.46.sav",
rawDataLocation = "[path to raw data on disk image]",
variableWorkbookFilename = "PRE.CISE_2021_variableManagement.xlsx",
variableWorkbookFileLocation = "[path to save location on Dropbox]"
)
}

}
\author{
Evelyn Yarzebinski, Computing Research Association
}
